task_status, task_category crud and add validation

+++ app/Http/Controllers/TaskController.php
@@ -2,12 +2,13 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\TaskStoreRequest;
+use App\Http\Requests\TaskUpdateRequest;
 use App\Models\Project;
 use App\Models\Task;
 use App\Models\TaskDeveloper;
 use App\Models\TaskReviewer;
 use App\Models\TaskStatus;
-use Illuminate\Http\Request;
 
 class TaskController extends Controller
 {
@@ -21,7 +22,7 @@ class TaskController extends Controller
     {
         return view('tasks.index', [
             'projectId' => $project->id,
-            'tasks' => Task::where('project_id', $project->id)->orderBy('id', 'desc')->paginate(15),
+            'tasks' => Task::with('task_status')->where('project_id', $project->id)->orderBy('id', 'desc')->paginate(15),
         ]);
     }
 
@@ -43,32 +44,53 @@ class TaskController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \App\Http\Requests\TaskStoreRequest  $request
      * @param  \App\Models\Project  $project
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request, Project $project)
+    public function store(TaskStoreRequest $request, Project $project)
     {
         $request->merge([
             'project_id' => $project->id,
         ]);
         $task = Task::create($request->all());
 
-        foreach ($request['developer_ids'] as $id) {
-            TaskDeveloper::create([
-                'task_id' => $task->id,
-                'user_id' => $id,
-            ]);
+        if ($request->developer_ids) {
+            foreach ($request->developer_ids as $id) {
+                TaskDeveloper::create([
+                    'task_id' => $task->id,
+                    'user_id' => $id,
+                ]);
+            }
         }
 
-        foreach ($request['reviewer_ids'] as $id) {
-            TaskReviewer::create([
-                'task_id' => $task->id,
-                'user_id' => $id,
-            ]);
+        if ($request->reviewer_ids) {
+            foreach ($request->reviewer_ids as $id) {
+                TaskReviewer::create([
+                    'task_id' => $task->id,
+                    'user_id' => $id,
+                ]);
+            }
         }
 
-        return redirect()->route('tasks.index', $project->id);
+        return redirect()
+            ->route('tasks.index', $project->id)
+            ->with('success', 'A task has been created.');
+    }
+
+    /**
+     * Display the specified resource.
+     *
+     * @param  \App\Models\Project  $project
+     * @param  \App\Models\Task  $task
+     * @return \Illuminate\Http\Response
+     */
+    public function show(Project $project, Task $task)
+    {
+        return view('tasks.detail', [
+            'projectId' => $project->id,
+            'taskDetail' => $task->detail,
+        ]);
     }
 
     /**
@@ -93,35 +115,41 @@ class TaskController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \App\Http\Requests\TaskUpdateRequest  $request
      * @param  \App\Models\Project  $project
      * @param  \App\Models\Task  $task
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Project $project, Task $task)
+    public function update(TaskUpdateRequest $request, Project $project, Task $task)
     {
         $request->merge([
             'project_id' => $project->id,
         ]);
         $task->update($request->all());
 
-        TaskDeveloper::where('task_id', $task->id)->delete();
-        foreach ($request['developer_ids'] as $id) {
-            TaskDeveloper::create([
-                'task_id' => $task->id,
-                'user_id' => $id,
-            ]);
+        if ($request->developer_ids) {
+            TaskDeveloper::where('task_id', $task->id)->delete();
+            foreach ($request->developer_ids as $id) {
+                TaskDeveloper::create([
+                    'task_id' => $task->id,
+                    'user_id' => $id,
+                ]);
+            }
         }
 
-        TaskReviewer::where('task_id', $task->id)->delete();
-        foreach ($request['reviewer_ids'] as $id) {
-            TaskReviewer::create([
-                'task_id' => $task->id,
-                'user_id' => $id,
-            ]);
+        if ($request->reviewer_ids) {
+            TaskReviewer::where('task_id', $task->id)->delete();
+            foreach ($request->reviewer_ids as $id) {
+                TaskReviewer::create([
+                    'task_id' => $task->id,
+                    'user_id' => $id,
+                ]);
+            }
         }
 
-        return redirect()->route('tasks.index', $project->id);
+        return redirect()
+            ->route('tasks.index', $project->id)
+            ->with('success', 'A task has been updated.');
     }
 
     /**
@@ -133,8 +161,10 @@ class TaskController extends Controller
      */
     public function destroy(Project $project, Task $task)
     {
-        Task::findOrFail($task->id)->delete();
+        $task->delete();
 
-        return redirect()->route('tasks.index', $project->id);
+        return redirect()
+            ->route('tasks.index', $project->id)
+            ->with('success', 'A task has been deleted.');
     }
 }
